+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed `game.log` file overfilling at the start of the game
+- Fixed RAM filling with the buffered game logs
+- Fixed Discord RPC updates
+
 ## [3.10.2] - 19.07.2024
 
 ### Added